requirements.txt
requests
beautifulsoup4
lxml
python-slugify
tqdm
tenacity